♻️ Mege RegisterSlogHandle and Wrap definitions to default.go
